refactor: moved sentry to own telemetry module and use store as source of truth
instead of module level variable that needs to be set separately
