+++ src/stores/ui.store.tsx
@@ -3,10 +3,12 @@ import { Assets } from "assets";
 import { Parser } from "expr-eval";
 import { solNative } from "lib/SolNative";
 import { CONSTANTS } from "lib/constants";
+import { defaultShortcuts } from "lib/shortcuts";
 import { googleTranslate } from "lib/translator";
+import MiniSearch from "minisearch";
 import {
-	autorun,
 	type IReactionDisposer,
+	autorun,
 	makeAutoObservable,
 	reaction,
 	runInAction,
@@ -20,11 +22,9 @@ import {
 	type NativeEventSubscription,
 } from "react-native";
 import type { IRootStore } from "store";
+import { safeCaptureException, setTelemetryEnabled as setTelemetryModuleEnabled } from "../telemetry";
 import { createBaseItems } from "./items";
-import MiniSearch from "minisearch";
-import { safeCaptureException, setTelemetryEnabled } from "../config";
 import { storage } from "./storage";
-import { defaultShortcuts } from "lib/shortcuts";
 
 const exprParser = new Parser();
 
@@ -243,8 +243,7 @@ export const createUIStore = (root: IRootStore) => {
 			);
 			solNative.setMediaKeyForwardingEnabled(store.mediaKeyForwardingEnabled);
 			solNative.setHyperKeyEnabled(store.hyperKeyEnabled);
-			setTelemetryEnabled(store.telemetryEnabled);
-			solNative.setTelemetryEnabled(store.telemetryEnabled);
+			store.setTelemetryEnabled(store.telemetryEnabled);
 			solNative.updateHotkeys(toJS(store.shortcuts));
 
 			store.username = solNative.userName();
@@ -452,8 +451,8 @@ export const createUIStore = (root: IRootStore) => {
 		},
 		setTelemetryEnabled: (enabled: boolean) => {
 			store.telemetryEnabled = enabled;
-			setTelemetryEnabled(enabled);
 			solNative.setTelemetryEnabled(enabled);
+			setTelemetryModuleEnabled(enabled);
 		},
 		rotateScratchPadColor: () => {
 			if (store.scratchPadColor === ScratchPadColor.SYSTEM) {
